Restructure tests layout to keep tests near the code being tested.
