+++ app/package.json
@@ -66,7 +66,7 @@
   },
   "jest": {
     "setupFiles": [
-      "./src/tests/setup.js"
+      "./jestsetup.js"
     ],
     "snapshotSerializers": [
       "<rootDir>/node_modules/enzyme-to-json/serializer"
