+++ app/src/scripts/faq/__tests__/faq.spec.jsx
@@ -1,4 +1,4 @@
-import Faq from '../../../scripts/faq/faq';
+import Faq from '../faq';
 
 describe('faq/faq/Faq', () => {
     it('renders successfully', () => {
