--- app/src/tests/.eslintrc
@@ -1,3 +1,0 @@
-env:
-    node: true
-    mocha: true
